Backend/requirements-deploy.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
bcrypt==4.3.0
cryptography==41.0.7
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
blinker==1.9.0
